leam-base.commons-jxpath.mid-1401.idx-37910.4.mutant

+++ commons-jxpath/src/main/java/org/apache/commons/jxpath/ri/model/beans/PropertyPointer.java
@@ -53,7 +53,7 @@ public abstract class PropertyPointer extends NodePointer {
      * @return int index
      */
     public int getPropertyIndex() {
-        return propertyIndex;
+        return (int)propertyIndex;
     }
 
     /**
